+++ src/gui/size.h
@@ -2,6 +2,7 @@
 #define ZC_GUI_SIZE_H
 
 #include <cstddef>
+#include <cstdint>
 
 namespace GUI
 {
@@ -15,39 +16,39 @@ public:
     template<typename T>
     static inline Size em(T t)
     {
-		return Size(static_cast<int>(t * static_cast<T>(emSize())));
+		return Size(static_cast<int32_t>(t * static_cast<T>(emSize())));
     }
 
     template<typename T>
     static inline Size largePixels(T t) noexcept
     {
-        return Size::sized(static_cast<int>(t));
+        return Size::sized(static_cast<int32_t>(t));
     }
 
     template<typename T>
     static inline Size smallPixels(T t) noexcept
     {
-        return Size::sized(static_cast<int>(t)*3/2);
+        return Size::sized(static_cast<int32_t>(t)*3/2);
     }
 
     template<typename T>
 	static inline constexpr Size pixels(T t) noexcept
 	{
-		return Size(static_cast<int>(t));
+		return Size(static_cast<int32_t>(t));
 	}
 
     /* Returns the size as the actual number of pixels. */
-	inline constexpr int resolve() const
+	inline constexpr int32_t resolve() const
 	{
 		return value;
 	}
 	
-	inline constexpr operator int() const noexcept
+	inline constexpr operator int32_t() const noexcept
 	{
 		return value;
 	}
 	
-	inline constexpr Size operator+(const int& other) const noexcept
+	inline constexpr Size operator+(const int32_t& other) const noexcept
 	{
 		return Size(value+other);
 	}
@@ -78,14 +79,14 @@ public:
 	}
 
 private:
-	int value;
+	int32_t value;
 
-	inline constexpr Size(int raw) noexcept: value(raw) {}
-    static int emSize();
-    static Size sized(int size) noexcept;
+	inline constexpr Size(int32_t raw) noexcept: value(raw) {}
+    static int32_t emSize();
+    static Size sized(int32_t size) noexcept;
 };
 
-inline Size operator ""_em(unsigned long long size)
+inline Size operator ""_em(uint64_t size)
 {
 	return Size::em(size);
 }
@@ -95,17 +96,17 @@ inline Size operator ""_em(long double size)
 	return Size::em(size);
 }
 
-inline constexpr Size operator ""_px(unsigned long long size)
+inline constexpr Size operator ""_px(uint64_t size)
 {
 	return Size::pixels(size);
 }
 
-inline Size operator ""_lpx(unsigned long long size)
+inline Size operator ""_lpx(uint64_t size)
 {
 	return Size::largePixels(size);
 }
 
-inline Size operator ""_spx(unsigned long long size)
+inline Size operator ""_spx(uint64_t size)
 {
 	return Size::smallPixels(size);
 }
